Ability to define the mtu globally (for all the container networks)
Currently, deployers can override MTUs on a per-network basis,
but may be looking for a way to define the mtu globally.

This implements a variable named ``lxc_container_default_mtu``
that currently defaults to 1500 (standard value). If this variable
is set, it will globally override the mtu, while still keeping the
per-network override.

Change-Id: Ia7e4f9e929f513ddc09e6de9467ec6ff765e2e6f
Closes-Bug: #1518311