Test container creation based on IP address
This test executes an LXC container listing from the LXC host, then
verifies that the container is present by name and IP address.

The test will work whether there is one or more IP addresses in
each containers created.

This patch also includes the following in order to unblock testing
for this role:
 - A forced apt-get update
 - A base other-requirements.txt file to ensure that the
   OpenStack-CI fallback requirements aren't used, which conflict
   with LXC and cause the instance to hang.

Change-Id: I2619b1090300a161c108fcc3f65059e8d50c851e