Fix container hostname for RFC 1034/1035
In lxc-container create this update will ensure we're using good
hostnames. This is required in some logging cases as well as SSL.
This will also allow a deployer to customize the domain name based on
their needs.

A new variable has been created called ``lxc_container_domain`` with a
default value of "openstack.local". This change has no upgrade impact as
any changes that need to happen within an existing environment will be
automatically taken care of without impacting the functionality of an
existing deployment.

This is largely based on Ib31a48dd480ecb376a6a8c5b35b09dfa5d2e58f6, with
the intent being to move the tasks from a playbook in the integrated
repo to the lxc_container_create role, and with the addition of limiting
127.0.0.1 to localhost within /etc/hosts.

Change-Id: I9e72699354a467478742281cab74690e66b6a652
Co-Authored-By: Kevin Carter <kevin.carter@rackspace.com>