Removed delegate_to when using with_first_found
The delegate_to function is unused when called with_first_found.
However in Ansible 2 using delegate_to and with_first_found forces
Ansible to login to the target host and search for files that are
nowhere to be found. this is an issue because Ansible 2 is more
strict on how tasks are executed.

Related-Bug: https://github.com/ansible/ansible/issues/14454

Change-Id: Ic06699d9ea47e692fe408e04ed2ebdc201ab3141
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>