Allow extra LXC config options via group/host vars
To reduce the number of times a container is restarted during the
build process, this patch implements the facility to consume
LXC container_config items from the inventory.

This allows the 'lxc.aa_profile' option, for example, to be set
in group_vars and therefore have it applied on container creation
instead of later when the service install playbook runs.

Change-Id: I4c95a231af3306c5010b168e7cc181deecf1a58d