Allow extra LXC command options via group/host vars
To reduce the number of times a container is restarted during the
build process, this patch implements the facility to consume
LXC container_command items from the inventory.

This allows any arbitrary command, such as the creation of a bind
mount directory, to be set in group_vars and therefore have it
applied on container creation instead of later when the service
install playbook runs.

Change-Id: I2aa015a136d6efca501797430d82f1461c303edf