Ensure that container config/command order is correct
This patch ensures that the config application and command execution
are completed in the same order as the lxc_container module.

This ensures that any bind-mounts to the host can have the container
directory created prior to the implementation of the drive mount.

With these tasks in the reverse order, the container cache would have
to have all the required directories in place already.

Change-Id: Ib8c289da2f4eb2d0e52b2d75376dd73dd9b2a650