Fix incorrect galera logs bind mount bug
The galera logs bind mount is broken in Mitaka and currently mounts
/var/log/galera to the host rather than /var/log/mysql_logs.

This patch removes the incorrect line from the LXC container config
file and adds the correct one using the lineinfile module. This
ensures that the task using lxc_container right after that will not
restart any galera containers.

Deployers will need to shut down / start their galera containers for
this fix to take effect. This is documented in the release notes.

Closes-Bug: 1609862
Change-Id: Ia129333e3aa10f337bb2f15c52c74685433a575d