Drop container setup script independent of container_networks
There is no reason there should be a loop or anything dependent on
container_networks being defined with this task as far as I can tell.

Since this task is skipped if container_networks is not set, the next
task will fail the role any time there are no container networks.

Change-Id: Id257fa3aa124b51f066d7ec0d4c30f4c8fb56e24