Use wait_for_connection instead of wait_for to check container
After configuring networking in the container, we used a task to wait
for SSH to respond on the container, using the Ansible wait_for module
trying to connect to port 22.

When managing containers in Ansible using other connection plugins that
use LXC commands on the physical host to execute modules, there may not
be an SSH daemon running within the container. In these situations we
can accomplish the same check without SSH by using the wait_for_connection
module (new in ansible 2.3) to verify that Ansible has a working execution
path into the container, but not necessarily via SSH.

Change-Id: I81beda2590b5f5505b7de9ae94cef06bcbdf2f93