Allow container creation based on variant property
Currently the variant used for container creation is a
universal override. This patch allows the use of an
inventory property to set the variant, allowing per
container overrides of the variant.

Also, a new variable is created which allows any
arbitrary additional options to be given to the creation
process, allowing other lxc-create options to be passed
in.

Change-Id: Id391c3ba245c53e1d82d7eb57187cbf1a952d6cf