Update instead of replacing the environment file
Currently the entire environment file is replaced,
which may remove any existing configuration on the
hosts including any pre-existing proxy configurations
and anything else the deployer may have put there.

This method replaces it with an additive process which
respects pre-existing content, but still allows the
global_environment_variables variable changes to be
fully reflected in the resulting environment file.

As the PATH setting in the previous template was
the default path which will already be present in
the container, that has been removed from the template.

Change-Id: I930f1711fbd56d2c97e8c80bd990350fa0c7ba73