Fix the bind mount options to match integrated repo
",create=dir" was added to the integrated repo in
I29be91b4dca5abe1bb8307429a86c06dbd3ef7aa to ensure that LXC handles
the bind mount directory creation inside the container.

Also a bug is fixed where bind mounts defined as part of the container
configuration list replace all '=' within the string as ' = '. To fix
this only the first = is padded with whitespace.

Change-Id: Ieed33a413ed9abec98e46cd519c18a95d5b6bd75