Fix logic for installing a routing file
In Ansible, a "string"|bool jinja2 filter does not result to a true
value as presented below:

~$ ansible localhost -e foo="string" -m debug -a "msg={{foo|bool}}"
localhost | SUCCESS => {
    "msg": false
}

Only a "true" string results in a boolean True value

~$ ansible localhost -e foo="true" -m debug -a "msg={{foo|bool}}"
localhost | SUCCESS => {
    "msg": true
}

as such, the routing files where never installed for the RedHat
distributions. We fix this logic by simply leaving this variable
undefined for the distributions who don't need it.

Change-Id: I0af0ad47487365ad593c5550d2c7e05f650e7174