Add variant local fact
In order to make it easier to detect the variant used
when a container was deployed, and therefore potentially
allowing smarter descisions for things like replacing
the container when a new variant is available, we
implement the variant as a local fact.

The file used to store facts will be the same for all
OpenStack services, with each service using its own section.

Example:

"ansible_local": {
    "openstack_ansible": {
        "lxc": {
            "variant": "default"
        }
    }
}

Change-Id: I340bfea36353381424f2abda62af814aabca98cf
(cherry picked from commit 64261a0c192f43a6f8338eba49acb46b91c31833)