templates: suse-interface: Never ask for DHCP configuration
There is no DHCP server running for that interface so we get errors
when no static IP is defined since wicked fails to obtain a lease for
this interface. We should either define a static IP if available or
leave it unconfigured. This is also in line with the ifcfg(5) manpage
suggestion to use BOOTPROTO=static without any IP configuration if
we just want to bring an interface up but without any further
configuration.

Change-Id: I4c1189def8fe21386d02e9ba9ddebbd8c44b55c7