handlers: main: Force kill a container on openSUSE if it takes too long to stop
It appears that lxc-1.1.5 is very slow when stopping a container. As a
result of which we pass the timeout (-t) option on openSUSE whilst this
is being investigated on the distribution level. This can also be
removed when lxc-2.X.X becomes the default option on openSUSE.

Link: https://bugzilla.opensuse.org/show_bug.cgi?id=1054609
Change-Id: I02299cb17477f63970e11ce3c8455cb12773a541
(cherry picked from commit 5209c772700d678bb4d1595cb20905c36bc15b15)