Add machinectl backend store options
This change modifies the lxc container create role to consume the
machinectl template when the `lxc_container_backing_store` is set to
"machinectl". This new container store type will allow us to better
manage containers and the images we use to run systems.

As mentioned in the release note within the lxc_hosts role, this change
has no upgrade impact as existing containers will continue to use the
storage they already have.

Change-Id: Id20b1b1eaeaad8ceb217ec82aac8053f9326094d
Depends-On: I70d53cabd0888954f31def924e9f4436398cdebf
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>