tasks: lxc_container_create: Check for 'lxc-info' failures and retry
If lxc-info command fails, it's best if we know it rather than ignoring
it as if nothing serious has happened. We expect the container to be
present at that point, so we should be able to retrieve its state
information. If lxc-info fails (perhaps the container hasn't been
fully created yet so we can't get its state) then we can simply retry
and fail if we can't recover in a reasonable amount of time.

Change-Id: Iefb03508138bc9dbafef8be9dc669c89cafc8916