Clean up container create task files
As it turns out, all the copy-on-write create task
files just include the cow task file so in this
patch we just clean it up to use symbolic links.

The cow task filename is also renamed to be
consistent with the others.

This also happens to fix a bug where the include
for the zfs file system referred to the wrong
file name for the include task.

Change-Id: I4ef6e2be526e18548334bd3ee9f759196501edeb