Fix container mac generation and remove extra tasks
The container create role will fail when adding a new network
to an existing set of container. The role attempts to read the hwaddr
file prior to its existence due to our use of local_facts. This change
simplifies how the mac addresses are generated by using a block / rescue
which will simply consume the known mac address from the container when
needed or fall back to the generation tasks as needed.

While working on this issue I noticed that we were still using a
pre-Ansible 2.0 failure notice which is no longer relevant as we've been
using Ansible 2.x for quite a while now. The old assertion has been
removed.

A new assertion has been added to notify the deployer that the
physical_host variable must be set to use this role.

Change-Id: Ic2800f1c17d10180e4e9a7be7f9b435ff8cc5487
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>