Fix container restart handler
The handler would try and stop a container before restarting it however
if the container was already stopped the handler would fail instead of
simply moving on to the next task. This change makes the "stop" portion
of the task detect the return status code of "2" when restarting the
container. If the return code is "2" we know that the container is
already stopped and that no change has occurred.

For the sake of consistency and to ensure the greatest chance for
success the test task that stops a container has also been given the
same setup.

Change-Id: Ia4856f36b2d106d987e3c774f31493e25a23d4b5
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>
(cherry picked from commit ff3441c90abbcb76a0f1b2fdf6d60b4ea0854cb2)