Generate a unique machine-id
The systemd machine-id needs to be unique on all network attached
devices. This change ensures that when a container comes online, a
unique machine-id is generated if one was not already present. When
the machine-id is created for the first time the container will restart
so the new ID can take effect.

More information on the machine-id can be found here:
https://www.freedesktop.org/software/systemd/man/machine-id.html

Change-Id: Ib25aeeecf1e6001e6c6b1a7d6b6d50eca7ab45fa
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>
(cherry picked from commit 01fc5fa64316993d5ae2eb9b37cc9afc2566cc5f)