Remove veth wiring check for machine-id
machine_id is not registered until further down in the file, so
this will fail with "The error was: |changed expects a dictionary"

We don't see the failure in our gates because the two preceding
conditions: not ((default_configuration_container | changed) or
           (bind_configuration_container | changed)

are always true, so the machine_id test is never used.

In an existing environment where the container is being updated
from an old configuration to the new networkd installation, it is
very possible that default_configuration_container and
bind_configuration_container are not changed, so the machine_id
var is checked for changed state. At that point ansible fails
because the var is undefined.

Change-Id: I0b95c6c5d0f52344d476e52219c1ce31edcf65da