Ensure container configuration is sorted
We convert the container networks from a dict to a list in order to
allow us to enumerate the networks and assign an ID which is required
for lxc > 3.

Since dicts are unsorted we should ensure that we sort the converted
list so that unnecessary container restarts are not initiated on
subsequent runs, where the id for the network could change due to the
initial dict being unsorted.

Change-Id: I6f86b32e05adb2e0e4e214d2eb4aa67cbc283be1