Implements possibility to use physical networks in containers.
This is usefull, if SR-IOV is used, and it's possible to
supply every container with it's own interface.
Especially usefull, if infinband is used (for example, for ceph),
as it can not be added into bridge.

openstack-user-config should be modified in this way for every service
within affected network:

- network:
    container_bridge: "ib5"
    container_type: "phys"
    container_interface: "ib5"
    ip_from_q: "storage"
    type: "raw"
    group_binds:
      - cinder_volume

Change-Id: I299aa3fb26f5278d2c86bb18c5df5abddcf0ff65
(cherry picked from commit 27697e9126cd32abf3e9dbb1eeac860805967edf)