Stop creating duplicate lines in lxc config files
The 'Ensure containers have access RO cgroups' will currently add the
same line to an lxc container's config file on every run of the role.

Add a regular expression to the task to check if the line to be inserted
already exists in the file, and if so, don't insert it again.

Change-Id: I5ac57bd59ea943c019dd0a5530c8246284b33269
(cherry picked from commit 296f428b9c7ab0c42cc609e5b63672df64456c38)