redhat: stop updating system unnecessarily
In RHEL-based operating systems, the update will actually attempt
to update every single package in the system which can result in
bad things (i.e. updating services when they're not supposed to).

The intention was to do the same as the apt-get update which simply
updates the repos. however this is not necessary in CentOS.

Change-Id: I5a1f86899406e125b3a790e2e85ce36d73fbc55e