Do not add container /etc/hosts entry for 127.0.1.1
This patch removes a task which adds an entry to /etc/hosts for
the container hostname / 127.0.1.1.

Some services, such as rabbitmq, require that the ip address they
are configured to listen on resolves to a hostname with the same
ip. Before this patch there were two /etc/hosts entries for the
hostname of each container, 127.0.1.1 and the container mgmt
network address.

This patch ensures that the /etc/hosts file is unambiguous and
always resolves the container hostname to the mgmt network ip
address.

Change-Id: I900bda8031d7b0798f73adf8985e64d4535be769