Replace linters test with integarted one
We've created integrated linters check job a while back and it's successfully
working for several releases. At the moment we experience difficulties
with future maintenance of the linters check from the openstack-ansible-tests
repo. So instead of fixing current one, we replace it with modern version of
the test.

We also replace regular functional tests with integrated scenario.

Change-Id: I4638a72507642a4a70aa09a9c9243c16d2bba376