Solving naming issue of OVS ports for lxc
There is a limitation of 15 characters for
ovs port names. So this role will use
the 15 ending characters of the container name
to generate port names(OVS and VNIC).
But veth-cleanup script will consider
this limitation only for OVS ports and not for the
virtual interface. So I added a condition to
also consider this limitation.
Note that the actual virtual interface has
the same name as OVS port and there is no problem
in the role.

Change-Id: I6a759c748469ce8c644727cb24bd4fa18cff3eda
(cherry picked from commit 297c0e17eec778f3b192503502e98f4188064099)